add player launch on alarm start

+++ src/utils/database.js
@@ -5,28 +5,32 @@
 const SQLite = require('react-native-sqlite-storage')
 
 const db = SQLite.openDatabase('db.db');
+const initTables = false;
 
-db.transaction((tx) => {
-    tx.executeSql('drop table if exists alarms');
-    tx.executeSql('drop table if exists playlists');
+if (initTables) {
+    db.transaction((tx) => {
+        tx.executeSql('drop table if exists alarms');
+        tx.executeSql('drop table if exists playlists');
 
-    tx.executeSql(
-        `create table if not exists alarms(
-            id INTEGER PRIMARY KEY AUTOINCREMENT,
-            time INTEGER,
-            repeat TEXT,
-            isEnabled INTEGER
-        );`
-    );
+        tx.executeSql(
+            `create table if not exists alarms(
+                id INTEGER PRIMARY KEY AUTOINCREMENT,
+                time INTEGER,
+                repeat TEXT,
+                isEnabled INTEGER
+            );`
+        );
 
-    tx.executeSql(
-        `create table if not exists playlists(
-            id INTEGER PRIMARY KEY AUTOINCREMENT,
-            name TEXT,
-            videos TEXT
-        );`
-    );
-});
+        tx.executeSql(
+            `create table if not exists playlists(
+                id INTEGER PRIMARY KEY AUTOINCREMENT,
+                name TEXT,
+                videos TEXT
+            );`
+        );
+    });
+}
+
 
 function makeDBCall(sql = '', args = []) {
     console.log('makeDBCall', sql, args);
